Module connect_four_gui::gui

Connect-K Client Side GUI

Builds the Client Side GUI of the ConnectK class.

Note that "cloning a Gtk-rs object only cost a pointer copy, so it’s not a problem." - GTK-rs documentation. This allows us to move data into closures to achieve our client-server functionality.

GTK+ is not thread-safe. Accordingly, none of this crate's structs implement Send or Sync.

Enums

Player

Functions

build_game_window
Builds game window with ConnectK board and play game button.

build_selection_game_window
Builds the parts of the GUI that allows user to select a game and create a game. Only games that can be joined are displayed.

connect_to_server
Connects to game server to get information about current, joinable games.

get_game
Gets game info from game server for specified game_id.

launch
Launches game GUI.

play_move
Plays the move specified by user. Returns the board (string) of the updated board.

poll_server
Sends requests to game server for specified game to determine whether opposing player has made a move.

update_board_gui
Updates the board GUI based on current board state.